refactor: remove explicit type arguments

+++ WatchingDirectoryTable/src/main/kotlin/example/App.kt
@@ -15,147 +15,147 @@ import javax.swing.table.DefaultTableModel
 import javax.swing.table.TableModel
 import javax.swing.table.TableRowSorter
 
-class MainPanel : JPanel(BorderLayout()) {
-  private val logger = JTextArea()
-  private val model = FileModel()
-  @Transient
-  private val sorter = TableRowSorter<FileModel>(model)
-  val deleteRowSet = mutableSetOf<Int>()
+private val logger = JTextArea()
+private val model = FileModel()
 
-  init {
-    val table = JTable(model)
-    table.setRowSorter(sorter)
-    table.setFillsViewportHeight(true)
-    table.setComponentPopupMenu(TablePopupMenu())
+@Transient
+private val sorter = TableRowSorter(model)
+val deleteRowSet = mutableSetOf<Int>()
 
-    val col = table.getColumnModel().getColumn(0)
-    col.setMinWidth(30)
-    col.setMaxWidth(30)
-    col.setResizable(false)
+fun makeUI(): Component {
+  val table = JTable(model)
+  table.rowSorter = sorter
+  table.fillsViewportHeight = true
+  table.componentPopupMenu = TablePopupMenu()
 
-    val dir = Paths.get(System.getProperty("java.io.tmpdir"))
-    val loop = Toolkit.getDefaultToolkit().getSystemEventQueue().createSecondaryLoop()
-    val worker = object : Thread() {
-      override fun run() {
-        runCatching {
-          FileSystems.getDefault().newWatchService()
-        }.onSuccess {
-          dir.register(it, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_DELETE)
-          append("register: $dir")
-          processEvents(dir, it)
-        }.onFailure {
-          append(it.message)
-        }
-        loop.exit()
-      }
-    }
-    worker.start()
-    if (!loop.enter()) {
-      append("Error")
-    }
+  val col = table.columnModel.getColumn(0)
+  col.minWidth = 30
+  col.maxWidth = 30
+  col.resizable = false
 
-    addHierarchyListener { e ->
-      val isDisplayableChanged = e.getChangeFlags().toInt() and HierarchyEvent.DISPLAYABILITY_CHANGED != 0
-      if (isDisplayableChanged && !e.getComponent().isDisplayable()) {
-        worker.interrupt()
-      }
-    }
-
-    val button = JButton("createTempFile")
-    button.addActionListener {
+  val dir = Paths.get(System.getProperty("java.io.tmpdir"))
+  val loop = Toolkit.getDefaultToolkit().systemEventQueue.createSecondaryLoop()
+  val worker = object : Thread() {
+    override fun run() {
       runCatching {
-        Files.createTempFile("_", ".tmp").toFile().deleteOnExit()
+        FileSystems.getDefault().newWatchService()
+      }.onSuccess {
+        dir.register(it, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_DELETE)
+        append("register: $dir")
+        processEvents(dir, it)
       }.onFailure {
         append(it.message)
       }
+      loop.exit()
     }
-
-    val p = JPanel()
-    p.add(button)
-
-    val sp = JSplitPane(JSplitPane.VERTICAL_SPLIT)
-    sp.setTopComponent(JScrollPane(table))
-    sp.setBottomComponent(JScrollPane(logger))
-    sp.setResizeWeight(.5)
-
-    add(p, BorderLayout.NORTH)
-    add(sp)
-    setPreferredSize(Dimension(320, 240))
+  }
+  worker.start()
+  if (!loop.enter()) {
+    append("Error")
   }
 
-  // Watching a Directory for Changes (The Java™ Tutorials > Essential Classes > Basic I/O)
-  // https://docs.oracle.com/javase/tutorial/essential/io/notification.html
-  // Process all events for keys queued to the watcher
-  fun processEvents(dir: Path, watcher: WatchService) {
-    while (true) {
-      // wait for key to be signaled
-      val key: WatchKey
-      try {
-        key = watcher.take()
-      } catch (ex: InterruptedException) {
-        EventQueue.invokeLater { append("Interrupted") }
-        return
-      }
-
-      for (event in key.pollEvents()) {
-        val kind = event.kind()
-
-        // This key is registered only for ENTRY_CREATE events,
-        // but an OVERFLOW event can occur regardless if events
-        // are lost or discarded.
-        if (kind === StandardWatchEventKinds.OVERFLOW) {
-          continue
-        }
-
-        // The filename is the context of the event.
-        @Suppress("UNCHECKED_CAST") val ev = event as WatchEvent<Path>
-        val filename = ev.context()
-
-        val child = dir.resolve(filename)
-        EventQueue.invokeLater {
-          append("$kind: $child")
-          updateTable(kind, child)
-        }
-      }
-
-      // Reset the key -- this step is critical if you want to
-      // receive further watch events.  If the key is no longer valid,
-      // the directory is inaccessible so exit the loop.
-      val valid = key.reset()
-      if (!valid) {
-        break
-      }
+  val button = JButton("createTempFile")
+  button.addActionListener {
+    runCatching {
+      Files.createTempFile("_", ".tmp").toFile().deleteOnExit()
+    }.onFailure {
+      append(it.message)
     }
   }
 
-  private fun updateTable(kind: WatchEvent.Kind<*>, child: Path) {
-    if (kind === StandardWatchEventKinds.ENTRY_CREATE) {
-      model.addPath(child)
-    } else if (kind === StandardWatchEventKinds.ENTRY_DELETE) {
-      for (i in 0 until model.getRowCount()) {
-        val path = model.getValueAt(i, 2)?.toString() ?: ""
-        if (path == child.toString()) {
-          deleteRowSet.add(i)
-          // model.removeRow(i);
-          break
-        }
-      }
-      sorter.setRowFilter(object : RowFilter<TableModel, Int>() {
-        override fun include(entry: Entry<out TableModel, out Int>) = !deleteRowSet.contains(entry.getIdentifier())
-      })
-    }
-  }
+  val p = JPanel()
+  p.add(button)
 
-  fun append(str: String?) {
-    logger.append(str + "\n")
+  val sp = JSplitPane(JSplitPane.VERTICAL_SPLIT)
+  sp.topComponent = JScrollPane(table)
+  sp.bottomComponent = JScrollPane(logger)
+  sp.resizeWeight = .5
+
+  return JPanel(BorderLayout()).also {
+    it.addHierarchyListener { e ->
+      val isDisplayableChanged = e.changeFlags.toInt() and HierarchyEvent.DISPLAYABILITY_CHANGED != 0
+      if (isDisplayableChanged && !e.component.isDisplayable) {
+        worker.interrupt()
+      }
+    }
+    it.add(p, BorderLayout.NORTH)
+    it.add(sp)
+    it.preferredSize = Dimension(320, 240)
   }
 }
 
-class FileModel : DefaultTableModel() {
+// Watching a Directory for Changes (The Java? Tutorials > Essential Classes > Basic I/O)
+// https://docs.oracle.com/javase/tutorial/essential/io/notification.html
+// Process all events for keys queued to the watcher
+private fun processEvents(dir: Path, watcher: WatchService) {
+  while (true) {
+    // wait for key to be signaled
+    val key: WatchKey
+    try {
+      key = watcher.take()
+    } catch (ex: InterruptedException) {
+      EventQueue.invokeLater { append("Interrupted") }
+      return
+    }
+
+    for (event in key.pollEvents()) {
+      val kind = event.kind()
+
+      // This key is registered only for ENTRY_CREATE events,
+      // but an OVERFLOW event can occur regardless if events
+      // are lost or discarded.
+      if (kind === StandardWatchEventKinds.OVERFLOW) {
+        continue
+      }
+
+      // The filename is the context of the event.
+      @Suppress("UNCHECKED_CAST") val ev = event as WatchEvent<Path>
+      val filename = ev.context()
+
+      val child = dir.resolve(filename)
+      EventQueue.invokeLater {
+        append("$kind: $child")
+        updateTable(kind, child)
+      }
+    }
+
+    // Reset the key -- this step is critical if you want to
+    // receive further watch events.  If the key is no longer valid,
+    // the directory is inaccessible so exit the loop.
+    val valid = key.reset()
+    if (!valid) {
+      break
+    }
+  }
+}
+
+private fun updateTable(kind: WatchEvent.Kind<*>, child: Path) {
+  if (kind === StandardWatchEventKinds.ENTRY_CREATE) {
+    model.addPath(child)
+  } else if (kind === StandardWatchEventKinds.ENTRY_DELETE) {
+    for (i in 0 until model.rowCount) {
+      val path = model.getValueAt(i, 2)?.toString() ?: ""
+      if (path == child.toString()) {
+        deleteRowSet.add(i)
+        // model.removeRow(i);
+        break
+      }
+    }
+    sorter.rowFilter = object : RowFilter<TableModel, Int>() {
+      override fun include(entry: Entry<out TableModel, out Int>) = !deleteRowSet.contains(entry.identifier)
+    }
+  }
+}
+
+fun append(str: String?) {
+  logger.append(str + "\n")
+}
+
+private class FileModel : DefaultTableModel() {
   private var number = 0
 
   fun addPath(path: Path) {
-    val obj = arrayOf(number, path.getFileName(), path.toAbsolutePath())
+    val obj = arrayOf(number, path.fileName, path.toAbsolutePath())
     super.addRow(obj)
     number++
   }
@@ -174,25 +174,28 @@ class FileModel : DefaultTableModel() {
     private val COLUMN_ARRAY = arrayOf(
       ColumnContext("No.", Integer::class.java, false),
       ColumnContext("Name", String::class.java, false),
-      ColumnContext("Full Path", String::class.java, false))
+      ColumnContext("Full Path", String::class.java, false)
+    )
   }
 }
 
-class TablePopupMenu : JPopupMenu() {
+private class TablePopupMenu : JPopupMenu() {
   private val delete = add("delete")
 
   init {
     delete.addActionListener {
-      val table = getInvoker() as? JTable ?: return@addActionListener
-      val model = table.getModel() as? DefaultTableModel ?: return@addActionListener
-      val selection = table.getSelectedRows()
-      for (i in selection.indices.reversed()) {
-        val idx = table.convertRowIndexToModel(selection[i])
-        model.getValueAt(idx, 2)?.toString()?.also {
-          runCatching {
-            Files.delete(Paths.get(it))
-          }.onFailure {
-            Toolkit.getDefaultToolkit().beep()
+      val table = invoker as? JTable
+      val model = table?.model as? DefaultTableModel
+      if (model != null) {
+        val selection = table.selectedRows
+        for (i in selection.indices.reversed()) {
+          val idx = table.convertRowIndexToModel(selection[i])
+          model.getValueAt(idx, 2)?.toString()?.also {
+            runCatching {
+              Files.delete(Paths.get(it))
+            }.onFailure {
+              Toolkit.getDefaultToolkit().beep()
+            }
           }
         }
       }
@@ -201,7 +204,7 @@ class TablePopupMenu : JPopupMenu() {
 
   override fun show(c: Component, x: Int, y: Int) {
     val table = c as? JTable ?: return
-    delete.setEnabled(table.getSelectedRowCount() > 0)
+    delete.isEnabled = table.selectedRowCount > 0
     super.show(c, x, y)
   }
 }
@@ -216,7 +219,7 @@ fun main() {
     }
     JFrame().apply {
       defaultCloseOperation = WindowConstants.EXIT_ON_CLOSE
-      contentPane.add(MainPanel())
+      contentPane.add(makeUI())
       pack()
       setLocationRelativeTo(null)
       isVisible = true
